+++ lib/src/storage_access_framework/file_types.dart
@@ -1,4 +1,5 @@
 /// To get the types for File statically
+@Deprecated('Only used by LegacySaf; will be removed in 3.0.0')
 class FileTypes {
   static const media = "media";
   static const image = "image";